Move all the agent stuff to `mod agent`.
